[WIP] moved all drawing calls to map service
